+++ packages/headless/src/features/facets/facet-set/facet-set-slice.test.ts
@@ -8,6 +8,7 @@ import {
   updateFacetNumberOfValues,
   updateFacetIsFieldExpanded,
   updateFreezeCurrentValues,
+  FacetRegistrationOptions,
 } from './facet-set-actions';
 import {buildMockFacetValue} from '../../../test/mock-facet-value';
 import {buildMockSearch} from '../../../test/mock-search';
@@ -16,7 +17,6 @@ import {executeSearch} from '../../search/search-actions';
 import {logSearchEvent} from '../../analytics/analytics-actions';
 import {buildMockFacetValueRequest} from '../../../test/mock-facet-value-request';
 import {buildMockFacetSearchResult} from '../../../test/mock-facet-search-result';
-import {FacetRegistrationOptions} from './interfaces/options';
 import {FacetResponse} from './interfaces/response';
 import {buildMockFacetRequest} from '../../../test/mock-facet-request';
 import {selectFacetSearchResult} from '../facet-search-set/specific/specific-facet-search-actions';
